fix: rebase issues, mobile responsiveness, upgrade to motion/react
