+++ packages/frontend/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@carely/core": "^1.0.0",
     "@hookform/resolvers": "^5.1.1",
+    "@radix-ui/react-hover-card": "^1.1.14",
     "@radix-ui/react-label": "^2.1.7",
     "@radix-ui/react-slot": "^1.2.3",
     "@tabler/icons-react": "^3.34.0",
@@ -20,8 +21,8 @@
     "axios": "^1.10.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "framer-motion": "^12.18.1",
     "lucide-react": "^0.515.0",
+    "motion": "^12.22.0",
     "next-themes": "^0.4.6",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
@@ -32,8 +33,7 @@
     "tailwind-merge": "^3.3.1",
     "tailwindcss": "^4.1.10",
     "vite": "^6.3.5",
-    "zod": "^3.25.64",
-    "@radix-ui/react-hover-card": "^1.1.14"
+    "zod": "^3.25.64"
   },
   "devDependencies": {
     "@eslint/js": "^9.25.0",
