--- packages/frontend/src/App.tsx
@@ -1,153 +1,0 @@
-import { SignedOut, SignedIn, UserButton } from "@clerk/clerk-react";
-import LandingPage from "./landing/landing.page";
-import { Navigate } from "react-router";
-import {
-  Navbar,
-  NavBody,
-  NavItems,
-  MobileNav,
-  NavbarLogo,
-  NavbarButton,
-  MobileNavHeader,
-  MobileNavToggle,
-  MobileNavMenu,
-} from "./components/ui/resizable-navbar";
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-
-function App() {
-  const navigate = useNavigate();
-  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
-
-  const navItems = [
-    {
-      name: "Features",
-      link: "#features",
-    },
-    {
-      name: "Contact",
-      link: "#contact",
-    },
-  ];
-
-  return (
-    <>
-      <SignedOut>
-        <div className="relative w-full">
-          <Navbar>
-            {/* Desktop Navigation */}
-            <NavBody>
-              <NavbarLogo />
-              <NavItems items={navItems} />
-              <div className="flex items-center gap-4">
-                <NavbarButton
-                  variant="secondary"
-                  onClick={() => navigate("/login")}
-                >
-                  Login
-                </NavbarButton>
-                <NavbarButton
-                  variant="primary"
-                  onClick={() => navigate("/login")}
-                >
-                  Get Started
-                </NavbarButton>
-              </div>
-            </NavBody>
-
-            {/* Mobile Navigation */}
-            <MobileNav>
-              <MobileNavHeader>
-                <NavbarLogo />
-                <MobileNavToggle
-                  isOpen={isMobileMenuOpen}
-                  onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
-                />
-              </MobileNavHeader>
-
-              <MobileNavMenu isOpen={isMobileMenuOpen}>
-                {navItems.map((item, idx) => (
-                  <a
-                    key={`mobile-link-${idx}`}
-                    href={item.link}
-                    onClick={() => setIsMobileMenuOpen(false)}
-                    className="relative text-neutral-600 dark:text-neutral-300"
-                  >
-                    <span className="block">{item.name}</span>
-                  </a>
-                ))}
-                <div className="flex w-full flex-col gap-4">
-                  <NavbarButton
-                    onClick={() => {
-                      setIsMobileMenuOpen(false);
-                      navigate("/login");
-                    }}
-                    variant="primary"
-                    className="w-full"
-                  >
-                    Login
-                  </NavbarButton>
-                  <NavbarButton
-                    onClick={() => {
-                      setIsMobileMenuOpen(false);
-                      navigate("/login");
-                    }}
-                    variant="primary"
-                    className="w-full"
-                  >
-                    Get Started
-                  </NavbarButton>
-                </div>
-              </MobileNavMenu>
-            </MobileNav>
-          </Navbar>
-        </div>
-        <LandingPage />
-      </SignedOut>
-      <SignedIn>
-        <div className="relative w-full">
-          <Navbar>
-            {/* Desktop Navigation */}
-            <NavBody>
-              <NavbarLogo />
-              <NavItems items={navItems} />
-              <div className="flex items-center gap-4">
-                <UserButton afterSignOutUrl="/" />
-              </div>
-            </NavBody>
-
-            {/* Mobile Navigation */}
-            <MobileNav>
-              <MobileNavHeader>
-                <NavbarLogo />
-                <div className="flex items-center gap-4">
-                  <UserButton afterSignOutUrl="/" />
-                  <MobileNavToggle
-                    isOpen={isMobileMenuOpen}
-                    onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
-                  />
-                </div>
-              </MobileNavHeader>
-
-              <MobileNavMenu isOpen={isMobileMenuOpen}>
-                {navItems.map((item, idx) => (
-                  <a
-                    key={`mobile-link-${idx}`}
-                    href={item.link}
-                    onClick={() => setIsMobileMenuOpen(false)}
-                    className="relative text-neutral-600 dark:text-neutral-300"
-                  >
-                    <span className="block">{item.name}</span>
-                  </a>
-                ))}
-              </MobileNavMenu>
-            </MobileNav>
-          </Navbar>
-        </div>
-        <Navigate to="/dashboard" />
-      </SignedIn>
-    </>
-  );
-}
-
-export default App;
